requirements.txt
langchain
pypdf
replicate
streamlit
python-dotenv
pandas
openai
xlsxwriter